Install test apps properly
