+++ dev/io.openliberty.transaction.remote_fat/test-applications/remoteEJB/src/com/ibm/ws/remoteEJB/localClient/LocalEJBClient.java
@@ -10,9 +10,9 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.remoteEJB.web;
+package com.ibm.ws.remoteEJB.localClient;
 
-import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 import javax.naming.InitialContext;
 import javax.naming.NamingException;
@@ -25,23 +25,21 @@ import javax.transaction.SystemException;
 
 import org.junit.Test;
 
+import com.ibm.ws.remoteEJB.client.EJBClient;
 import com.ibm.ws.remoteEJB.shared.TestBeanRemote;
 
 @SuppressWarnings("serial")
-@WebServlet("/RemoteEJBClient")
-public class RemoteEJBClient extends EJBClient {
+@WebServlet("/LocalEJBClient")
+public class LocalEJBClient extends EJBClient {
 
     @Override
     public void init() {
         System.getProperties().entrySet().stream().forEach(e -> System.out.println("Prop: " + e.getKey() + " -> " + e.getValue()));
         System.getenv().entrySet().stream().forEach(e -> System.out.println("Env: " + e.getKey() + " -> " + e.getValue()));
 
-        final int remoteIIOPPort = Integer.getInteger("bvt.prop.IIOP.secondary");
-
         try {
             Object found = new InitialContext()
-                            .lookup("corbaname:iiop:localhost:" + remoteIIOPPort
-                                    + "/NameService#ejb/global/TestBeanApp/TestBeanEJB/TestBean!com.ibm.ws.remoteEJB.shared.TestBeanRemote");
+                            .lookup("com.ibm.ws.remoteEJB.shared.TestBeanRemote");
             bean = (TestBeanRemote) PortableRemoteObject.narrow(found, TestBeanRemote.class);
         } catch (NamingException e) {
             e.printStackTrace();
@@ -49,24 +47,8 @@ public class RemoteEJBClient extends EJBClient {
     }
 
     @Test
-    public void testBeanIsRemote(HttpServletRequest request,
-                                 HttpServletResponse response) throws NotSupportedException, SystemException, NamingException {
-        assertFalse("Bean is running locally", System.getProperty(SERVER_NAME_PROPERTY).equals(bean.getProperty(SERVER_NAME_PROPERTY)));
-    }
-
-    /* The methods below override the tests that will fail until propagation is implemented */
-    @Override
-    public void testRequiredWith(HttpServletRequest request,
-                                 HttpServletResponse response) throws NotSupportedException, SystemException, NamingException {
-    }
-
-    @Override
-    public void testMandatoryWith(HttpServletRequest request,
-                                  HttpServletResponse response) throws NotSupportedException, SystemException, NamingException {
-    }
-
-    @Override
-    public void testSupportsWith(HttpServletRequest request,
-                                 HttpServletResponse response) throws NotSupportedException, SystemException, NamingException {
+    public void testBeanIsLocal(HttpServletRequest request,
+                                HttpServletResponse response) throws NotSupportedException, SystemException, NamingException {
+        assertTrue("Bean is running remotely", System.getProperty(SERVER_NAME_PROPERTY).equals(bean.getProperty(SERVER_NAME_PROPERTY)));
     }
 }
